+++ R/helpers.R
@@ -21,42 +21,6 @@ save_plot <- function(gg_plot, ...) {
   )
 }
 
-#' Count Specimens
-#'
-#' Tabulate the total number of distinct specimen collections by collector,
-#' colleciton number and date while accounting for herbarium duplicates.
-#'
-#' @param spp_tibble Specimen tibble to count unique specimen vouchers.
-#' @export
-#'
-#' @return Numeric scalar with the number of distinct specimen vouchers.
-#'
-#' @examples
-#' count_specimens(spp_tibble = herbarium_specimens)
-#'
-count_specimens <- function(spp_tibble) {
-
-  # Calculate total number of specimens accounting for duplicated records.
-  specimen_count <- spp_tibble %>%
-    dplyr::select("Collector", "Collection_Number", "Date", "Herbarium") %>%
-    dplyr::mutate(
-      row_id = 1:nrow(.),
-      Herbarium = stringr::str_remove_all(
-        string = .data$Herbarium,
-        pattern = "\\[|\\]"
-      )  %>%
-        stringr::str_split(string = ., pattern = ", +") %>%
-        purrr::map_chr(., function(herbarium) {
-          herbarium_split <- unlist(herbarium) %>% sort() %>%
-            stringr::str_c(., collapse = " ")
-          ifelse(length(herbarium_split) == 1, herbarium_split, "NA")
-          })
-      ) %>%
-    dplyr::distinct(., .data$Collector, .data$Collection_Number,
-                    .data$Date, .data$Herbarium, .keep_all = TRUE) %>% nrow()
-  return(specimen_count)
-}
-
 #' Capitalize String Character
 #'
 #' Given an input character vector, capitalize the first element character.
@@ -96,51 +60,3 @@ italicize <- function(unitalicized, chunk_type) {
     paste0("*", unitalicized, "*")
   }
 }
-
-#' Filter Reviewed Specimens
-#'
-#' Subset specimen tibble to exclude outgroups, leaving only species of interest.
-#'
-#' @details
-#' Reviewed species include:
-#' - Physaria
-#' - P. acutifolia
-#' - P. brassicoides
-#' - P. condensata
-#' - P. didymocarpa subsp. didymocarpa
-#' - P. didymocarpa subsp. lanata
-#' - P. didymocarpa subsp. lyrata
-#' - P. dornii
-#' - P. eburniflora
-#' - P. integrifolia
-#' - P. 'medicinae'
-#' - P. vitulifera
-#'
-#' @inheritParams layer_specimens
-#' @importFrom dplyr filter select add_count distinct
-#' @export
-#'
-#' @return Tibble of specimens filtered by species of reviewed annotation.
-#'
-#' @examples
-#' herbarium_specimens %>%
-#'   filter_reviewed(specimen_tbl = .) %>%
-#'   dplyr::select(Taxon_a_posteriori) %>%
-#'   dplyr::add_count(Taxon_a_posteriori) %>%
-#'   dplyr::distinct()
-#'
-filter_reviewed <- function(specimen_tbl) {
-  filtered_specimens <- specimen_tbl %>%
-    filter(
-      !grepl(
-        pattern = "outgroup|excluded",
-        x = .data$excel_sheet
-      ),
-      !grepl(
-        pattern = paste("chambersii", "floribunda", "rollinsii", sep = "|"),
-        x = .data$Taxon_a_posteriori
-      )
-    )
-  return(filtered_specimens)
-}
-